Moving transfer timeout test to test_llm_pytorch, fixing broken py_executor.py
